fix: Add `FromJSON1` and `ToJSON1` instances.
We need these for `toJSON` to work on `Fix` types.
