+++ BUILD.bazel
@@ -83,7 +83,7 @@ haskell_library(
         ],
     ),
     src_strip_prefix = "src",
-    version = "0.0.6",
+    version = "0.0.7",
     visibility = ["//visibility:public"],
     deps = [
         ":lexer",
@@ -100,6 +100,7 @@ haskell_library(
         hazel_library("groom"),
         hazel_library("mtl"),
         hazel_library("text"),
+        hazel_library("transformers-compat"),
     ],
 )
