+++ cimple.cabal
@@ -1,5 +1,5 @@
 name:                 cimple
-version:              0.0.6
+version:              0.0.7
 synopsis:             Simple C-like programming language
 homepage:             https://toktok.github.io/
 license:              GPL-3
@@ -32,7 +32,8 @@ library
     , Language.Cimple.Pretty
     , Language.Cimple.Program
   other-modules:
-      Language.Cimple.AST
+      Language.Cimple.Annot
+    , Language.Cimple.AST
     , Language.Cimple.Flatten
     , Language.Cimple.Graph
     , Language.Cimple.Lexer
